==> methodology_immigration/initial/1.4_Download_all_related_tweets_from_local_authors.ipynb <==
## Downloading all the related tweets from our final list of authors.

Now that we cleaned our authors lefting mostly chilean ones, we will go to download all the tweets from this authors that are related with our topic.

import pandas as pd
import pickle

**Manual input:** Set the path.

path='C:/Users/acoub/OneDrive/Desktop/DSDM/Thesis/ChileGov/methodology_immigration'
sep='/'

path_scripts = path+sep+'initial'
path_data= path+sep+'data'

file = path_data+sep+'keywords.sav'
list_of_words=pickle.load(open(file, 'rb'))

filename = path_data+sep+'start_search_date.sav'
start_date= pickle.load(open(filename, 'rb'))

filename = path_data+sep+'end_search_date.sav'
end_date= pickle.load(open(filename, 'rb'))

file = path_data+sep+'filtered_by_location.sav'
df = pickle.load(open(file, 'rb'))

Here we obtain the final list of authors to look for all the topic related tweets from these authors.

authors=list(df['author.username'])
authors = list(set(authors))

query=""
for word in list_of_words:
    query+= word + " OR "
query=query[:len(query)-4]

query="("+query+")" + "-is:retweet"

We create a list of queries with the keywords and the authors. Considering that the maximum lenght of a query is 1,024 characters, we should create a list of queries, each one with less than 1,024 characters. After, with the option searches we run from a txt file all the queries.

query_text= query
queries_list=[]
query_from_good=""
query_from_test=""

for author in authors:
    query_from_test+="from: " + author + " OR "
    if len(query_text + query_from_test) + 3 <= 1024:
        query_from_good = query_from_test
    else:
        queries_list.append("("+query_from_good[:len(query_from_good)-4]+")"+ query_text)
        query_from_good="from: " + author + " OR "
        query_from_test="from: " + author + " OR "

queries_list[0]

len(queries_list)

with open(path_data+sep+'final_search_for_tweets_df.txt', 'w') as f:
    for line in queries_list:
        f.write(line)
        f.write('\n')

**Manual revision:** Run the following two cells to know the total number of tweets to download and see if does not exceed the limit of your twarc acount. If is OK run also the next cell.

file=path_data+sep+'final_search_for_tweets_df.txt'
%cd {path_data}
! twarc2 searches --archive --counts-only --granularity day --start-time {start_date} --end-time {end_date} {file} final_search_count.csv

count=pd.read_csv(path_data+sep+'final_search_count.csv',encoding = "ISO-8859-1")
print("Tweets to download: "+str(count['day_count'].sum()))

**Download the tweets:** (Only if you have enought limit in your twarc account)

%cd {path_data}
! twarc2 searches --archive --start-time {start_date} --end-time {end_date} {file} final_corpus_search.json
