==> methodology_immigration/initial/2.2_Downloading_Left_Right_Hashtags_To_Label.ipynb <==
import pandas as pd
import pickle

**Manual Input**: Set the path.

path='C:/Users/acoub/OneDrive/Desktop/DSDM/Thesis/ChileGov/methodology_immigration'
sep='/'

path_scripts = path+sep+'initial'
path_data= path+sep+'data'

file = path_data+sep+'final_data_set.sav'
df = pickle.load(open(file, 'rb'))

file = path_data+sep+'left_hashtags_for_label.sav'
left_hashtags = pickle.load(open(file, 'rb'))

file = path_data+sep+'right_hashtags_for_label.sav'
right_hashtags= pickle.load(open(file, 'rb'))

file = path_data+sep+'start_date_for_label.sav'
start_date= pickle.load(open(file, 'rb'))

file = path_data+sep+'end_date_for_label.sav'
end_date= pickle.load(open(file, 'rb'))

hashtag_string="("
for hashtag in (left_hashtags+right_hashtags):
    hashtag_string+=hashtag + " OR "
hashtag_string=hashtag_string[:len(hashtag_string)-4]+")"
    
authors=list(df['author.username'])
authors = list(set(authors))

query_text= hashtag_string
queries_list=[]
query_from_good=""
query_from_test=""

for author in authors:
    query_from_test+="from: " + author + " OR "
    if len(query_text + query_from_test) + 3 <= 1024:
        query_from_good = query_from_test
    else:
        queries_list.append("("+query_from_good[:len(query_from_good)-4]+")"+ query_text)
        query_from_good="from: " + author + " OR "
        query_from_test="from: " + author + " OR "

print(len(queries_list))

queries_list[0]

with open(path_data+sep+'search_for_labels.txt', 'w') as f:
    for line in queries_list:
        f.write(line)
        f.write('\n')

**Manual revision:** Run the following two cells to know the total number of tweets to download and see if does not exceed the limit of your twarc acount. If is OK run also the next cell.

file=path_data+sep+'search_for_labels.txt'
%cd {path_data}
! twarc2 searches --archive --counts-only --granularity day --start-time {start_date} --end-time {end_date} {file} labels_count.csv

count=pd.read_csv(path_data+sep+'labels_count.csv',encoding = "ISO-8859-1")
print("Tweets to download: "+str(count['day_count'].sum()))

**Download the tweets:** (Only if you have enought limit in your twarc account)

%cd {path_data}
! twarc2 searches --archive --start-time {start_date} --end-time {end_date} {file} right_left_hashtags.json
